created class for parsing questions
