+++ src/quiz/entities/Quiz.java
@@ -1,22 +1,14 @@
 package quiz.entities;
 
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
+import quiz.service.QuestionParser;
+
 import java.text.MessageFormat;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Scanner;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
@@ -29,10 +21,6 @@ import static java.util.ResourceBundle.getBundle;
  * @author Javatlacati
  */
 public class Quiz {
-
-    public static final String FIELD_DIVIDER = "@@";
-    private static final Pattern SEPARATE_BY_DIVIDER_PATTERN = Pattern.compile(FIELD_DIVIDER);
-    private static final Pattern LINE_BREAK = Pattern.compile("\\\\n");
     private static final Pattern BLANKS = Pattern.compile("\\s+");
 
     private static final Scanner scanner = new Scanner(System.in);
@@ -225,75 +213,6 @@ public class Quiz {
     }
 
     /**
-     * Adds a new multiple choice question to the specified question list.
-     */
-    private void addMultipleChoiceQuestion(final List<Question> questions,
-                                           final String vettedness, final String explanation, final String questionText, final int correctAnswerIdx,
-                                           final String category,
-                                           final String difficulty,
-                                           final String... answersTexts) {
-        final MultipleChoiceQuestion choiceQuestion = new MultipleChoiceQuestion(vettedness);
-        choiceQuestion.setExplanation(explanation);
-        choiceQuestion.setText(questionText);
-        choiceQuestion.setCategory(category);
-        choiceQuestion.setDifficulty(str2difficulty(difficulty));
-        for (int i = 0; i < answersTexts.length; i++) {
-            final String answerText = answersTexts[i];
-            choiceQuestion.setChoice(answerText, i == correctAnswerIdx);
-        }
-        questions.add(choiceQuestion);
-    }
-
-    /**
-     *
-     */
-    private void addMultipleAnswerQuestion(final List<Question> questions,
-                                           final String vettedness, final String explanation, final String questionText,
-                                           final String category, final String difficulty,
-                                           final Map<String, Boolean> answerChoicesMap) {
-        final MultipleAnswerQuestion choiceQuestion = new MultipleAnswerQuestion(vettedness);
-        choiceQuestion.setExplanation(explanation);
-        choiceQuestion.setText(questionText);
-        choiceQuestion.setCategory(category);
-        choiceQuestion.setDifficulty(str2difficulty(difficulty));
-        answerChoicesMap.forEach(choiceQuestion::setChoice);
-        questions.add(choiceQuestion);
-    }
-
-    /**
-     *
-     */
-    private void addFillBlankQuestion(final List<Question> questions,
-                                      final String vettedness, final String explanation, final String questionText,
-                                      final String category, final String difficulty,
-                                      final String... blanks) {
-        final FillBlankQuestion fillBlankQuestion = new FillBlankQuestion(vettedness);
-        fillBlankQuestion.setExplanation(explanation);
-        fillBlankQuestion.setText(questionText);
-        fillBlankQuestion.setCategory(category);
-        fillBlankQuestion.setDifficulty(str2difficulty(difficulty));
-        for (final String blank : blanks) {
-            fillBlankQuestion.setAnswer(blank);
-        }
-        questions.add(fillBlankQuestion);
-    }
-
-    /**
-     *
-     */
-    private Map<String, Boolean> createAnswerChoicesMap(final String[] answerTexts, final Boolean... answerValidities) {
-        final int numsOfAnswers = answerTexts.length;
-        if (numsOfAnswers != answerValidities.length) {
-            throw new RuntimeException(MessageFormat.format(getBundle("quiz/resources/quiz").getString("ANSWER_TEXTS_ERR"), new Object[]{}));
-        }
-        final HashMap<String, Boolean> hashMap = new HashMap<>(2);
-        for (int answIdx = 0; answIdx < answerTexts.length; answIdx++) {
-            hashMap.put(answerTexts[answIdx], answerValidities[answIdx]);
-        }
-        return hashMap;
-    }
-
-    /**
      * Prints again failed questions showing also their correct answers
      */
     public void showFailed() {
@@ -312,67 +231,6 @@ public class Quiz {
                 System.out.println(MessageFormat.format(getBundle("quiz/resources/quiz").getString("EXPLANATION"), questions.get(i).getExplanation()));
             }
         }
-    }
-
-    /**
-     * @param questionString the question fields split by @@
-     */
-    public void parseQuestion(final String questionString) {
-        final String[] questarray = SEPARATE_BY_DIVIDER_PATTERN.split(questionString);
-        String tipoPregunta = questarray[0];
-        String vettedness = "v".equals(questarray[1]) ? Question.VETTED : Question.TRIAL;
-        String explanation = questarray[2];
-        String category = questarray[3];
-        String difficulty = questarray[4];
-        String questionText = formateaPregunta(questarray[5]);
-        switch (tipoPregunta) {
-            case "MC": {
-                //multiple choice
-                int correctAnswerIdx = Integer.parseInt(questarray[6]);
-                String[] answersTexts = Arrays.copyOfRange(questarray, 7, questarray.length);
-                addMultipleChoiceQuestion(questions, vettedness, explanation,
-                        questionText, correctAnswerIdx, category, difficulty, answersTexts);
-                break;
-            }
-            case "FB": {
-                //fill in the blanks
-                String[] blanks = Arrays.copyOfRange(questarray, 6, questarray.length);
-                addFillBlankQuestion(questions, vettedness, explanation, questionText, category, difficulty, blanks);
-                break;
-            }
-            case "MA": {
-                Map<String, Boolean> choices = parseChoicesMap(Arrays.copyOfRange(questarray, 6, questarray.length));
-                addMultipleAnswerQuestion(questions, vettedness, explanation, questionText,
-                        category, difficulty, choices);
-                break;
-            }
-            case "OE":{
-                //one example
-
-            }
-            default:
-                System.err.println(MessageFormat.format(getBundle("quiz/resources/quiz").getString("QUESTION_TYPE_ERR"), new Object[]{}));
-                break;
-        }
-
-    }
-
-    /**
-     * Parses the choices and relates them with their values.
-     *
-     * @param questionMapArr list with options and then values of each question
-     * @return options maped with their values
-     */
-    public Map<String, Boolean> parseChoicesMap(final String... questionMapArr) {
-        final String[] questarray = Arrays.copyOfRange(questionMapArr, 0, questionMapArr.length / 2);
-        final String[] strAnswerValidities = Arrays.copyOfRange(questionMapArr, questionMapArr.length / 2, questionMapArr.length);
-        final Boolean[] answerValidities = new Boolean[strAnswerValidities.length];
-        Arrays.stream(strAnswerValidities).map((booleanAnswer) -> "true".equalsIgnoreCase(booleanAnswer) ? Boolean.TRUE : Boolean.FALSE).collect(Collectors.toList()).toArray(answerValidities);
-        return createAnswerChoicesMap(questarray, answerValidities);
-    }
-
-    private static String formateaPregunta(final String strPregunta) {
-        return LINE_BREAK.matcher(strPregunta).replaceAll("\n");
     }
 
     public void askForSubsetSize() {
@@ -411,21 +269,13 @@ public class Quiz {
         System.out.println("Difficulty levels:");
         System.out.println("0 - normal\n 1 - hard\n 2 - easy");
         List<Difficulty> selectedDifficulties = Arrays.stream(BLANKS.split(scanner.nextLine())).distinct()
-                .map(String::trim).map(Quiz::str2difficulty).collect(Collectors.toList());
+                .map(String::trim).map(QuestionParser::str2difficulty).collect(Collectors.toList());
         System.out.println("You selected the categories:" + Arrays.toString(selectedDifficulties.toArray()));
         //filter questions by category
         questions = questions.stream().filter(question -> selectedDifficulties.contains(question.getDifficulty())).collect(Collectors.toList());
     }
 
-    private static Difficulty str2difficulty(String idxStr) {
-        switch (idxStr) {
-            case "2":
-                return Difficulty.EASY;
-            case "1":
-                return Difficulty.HARD;
-            case "0":
-            default:
-                return Difficulty.NORMAL;
-        }
+    public List<Question> getQuestions() {
+        return questions;
     }
 }
